refactor: extract PDF preview
